+++ src/Krasnoludki.Core/Graph/EdgeFlow.cs
@@ -6,11 +6,11 @@ public class EdgeFlow
     public int To{ get; }
     public int Capacity{ get; }
     public int CurrFlow{ get; private set;}
-    public double Cost { get; }
+    public long Cost { get; }
     public EdgeFlow BackwardEdge{ get; }
 
     
-    public EdgeFlow(int from, int to, int capacity, double cost = 0)
+    public EdgeFlow(int from, int to, int capacity, long cost = 0)
     {
         From = from;
         To = to;
